backend/requirements.txt
django>=5.1.1
channels>=4.0.0
djangorestframework>=3.14.0
python-dotenv>=1.0.0
python-binance
asyncio-throttle